refactor the get first image for gallery logic
Previously, the gallery view was doing a LINQ query to fetch the first
image of a gallery entry. But this would break when there was a gallery
entry that had no images. SO I refactored this into a getter property on
the GalleryEntry model, so that the view can simply call this getter
method and always expect a URL to be returned, regardless of whether or
not the gallery entry has any associated images yet or not